Cleaned the code
