+++ src/pages/MoviePage.jsx
@@ -1,57 +1,28 @@
 import styled from "styled-components";
-import { useMovieDetails } from "../hooks/fetch.js";
+import { useMovieDetails } from "../hooks/useMovieDetails.js";
 import { useParams } from "react-router-dom";
 import { BackLink } from "../components/BackLink.jsx";
 import { MoviePoster } from "../components/MoviePoster.jsx";
-import { Rating } from "../components/Rating.jsx";
-import { ComingSoonTag } from "../components/ComingSoonTag.jsx";
+import { MovieInfo } from "../components/MovieInfo.jsx";
 import { ErrorComponent } from "../components/ErrorComponent.jsx";
 
 const PosterBackground = styled.div`
-  background-image: ${({ poster }) => `
-  linear-gradient(rgba(0, 0, 0, 0) 70%, rgb(0, 0, 0) 100%),
-  url(https://image.tmdb.org/t/p/original${poster})
-`};
+  ${({ poster }) =>
+    poster
+      ? `
+    background-image: linear-gradient(rgba(0, 0, 0, 0) 70%, rgb(0, 0, 0) 100%),
+    url(https://image.tmdb.org/t/p/original${poster});
+    background-size: cover;
+    `
+      : `
+    background-color: black;
+    `}
   min-height: 100vh;
   display: flex;
   background-size: cover;
   flex-direction: column;
   justify-content: flex-end;
   color: white;
-`;
-
-const MovieWrapper = styled.div`
-  padding: 50px;
-  display: flex;
-  flex-direction: column;
-  margin-top: 24px;
-
-  @media (min-width: 577px) {
-    flex-direction: row;
-    align-items: flex-end;
-  }
-`;
-
-const MovieDetailsWrapper = styled.div`
-  max-width: 400px;
-  @media (min-width: 577px) {
-    margin: 0 0 0 20px;
-  }
-`;
-
-const StyledH1 = styled.h1`
-  font-size: 24px;
-  margin: 20px 0 0 0;
-  display: flex;
-  align-items: center;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 8px;
-`;
-
-const Title = styled.span`
-  margin: 0 10px 0 0;
-  text-shadow: 1px 1px #4d4d4d;
 `;
 
 export const MoviePage = () => {
@@ -63,27 +34,11 @@ export const MoviePage = () => {
   if (!details) return null;
 
   return (
-    <div>
-      <PosterBackground poster={details.backdrop_path || ""}>
+    <>
+      <PosterBackground poster={details.backdrop_path}>
         <BackLink />
-        <MovieWrapper>
-          <MoviePoster
-            src={`https://image.tmdb.org/t/p/w500${details.poster_path}`}
-            alt={`Movie Poster of ${details.title}`}
-          />
-          <MovieDetailsWrapper>
-            <StyledH1>
-              <Title>{details.title}</Title>
-              <Rating
-                value={details.vote_average}
-                releaseDate={details.release_date}
-              />
-            </StyledH1>
-            <ComingSoonTag releaseDate={details.release_date} />
-            <p>{details.overview}</p>
-          </MovieDetailsWrapper>
-        </MovieWrapper>
+        <MovieInfo details={details} />
       </PosterBackground>
-    </div>
+    </>
   );
 };
